requirements.txt
ipykernel
openai
pandas
matplotlib
seaborn
python-dotenv
yfinance
colorama
openai
